requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
gTTS
Pillow